switch the totalping property type to ushort

+++ src/DNSUtility.Service/Benchmarks/IBenchmark.cs
@@ -4,5 +4,5 @@ namespace DNSUtility.Service.Benchmarks;
 
 public interface IBenchmark
 {
-    long RunBenchmark(string address);
+    ushort RunBenchmark(string address);
 }
